Add "show video window" buttons
The Winamp image logo in the main window and the Winamp text logo in the small/focus window are now buttons which will show the video player window. This skin previously had no button to show/hide the video window, so if you closed it, there was no path to make it reappear.